requirements.txt

langchain
langchain-community
langchain-pinecone
PyPDF
langchain-openai 
tiktoken
pinecone-client

. e